Remove StyleResource and StyleBinding. Fold XamlBinding into Binding since that is the only place it is being used now.
